+++ DesignToolbox/DesignToolbox/Pages/Components/Layouts/ColoredSurface/ColoredSurfacePage.swift
@@ -44,13 +44,13 @@ private struct ColoredSurfaceDemo: View {
         OUDSColoredSurface(color: configurationModel.selectedColor.toSurfaceColor(from: theme)) {
             VStack(alignment: .center, spacing: theme.spaces.fixedMedium) {
                 Text(configurationModel.selectedColor.formattedName)
-                    .oudsForegroundColor(theme.colors.contentDefault)
+                    .foregroundColor(theme.colors.contentDefault)
 
-                OUDSButton(text: "app_components_button_tech".localized(),
+                OUDSButton("app_components_button_tech",
                            appearance: .default,
                            style: .default) {}
 
-                OUDSLink(text: "app_components_link_tech".localized(), indicator: .next) {}
+                OUDSLink("app_components_link_tech", indicator: .next) {}
             }
             .frame(maxWidth: .infinity, alignment: .center)
             .padding(.all, theme.spaces.fixedMedium)