+++ DesignToolbox/DesignToolbox/Pages/Components/Navigations/TabBar/TabBarElement.swift
@@ -30,10 +30,10 @@ struct TabBarElement: DesignToolboxElement {
     }
 }
 
-private struct TabBarIllustration: View {
+struct TabBarIllustration: View {
 
     var body: some View {
-        OUDSTabBar(selected: 0, count: 3) {
+        OUDSTabBar(selectedTab: .constant(0), count: 3) {
             let wording = "app_components_common_label_label".localized()
             FakeTabItem(title: wording, imageName: "heart-empty", tag: 0)
             FakeTabItem(title: wording, imageName: "heart-empty", tag: 1)