+++ DesignToolbox/DesignToolbox/Pages/Utils/Elements/DesignToolboxElementsPage.swift
@@ -33,12 +33,11 @@ struct DesignToolboxElementsPage: View {
     // MARK: Body
 
     var body: some View {
-        #if os(iOS) || os(visionOS)
-        NavigationView {
+        #if os(iOS)
+        OUDSNavigationStack {
             elementsPage
-                .navigationBarTitleDisplayMode(.inline)
         }
-        .navigationViewStyle(.stack)
+        .navigationBarTitleDisplayMode(.inline)
         #else // macOS
         // Trick to be sure the view refreshes because NavigationView not always refreshed with macOS
         NavigationSplitView {
@@ -63,7 +62,7 @@ struct DesignToolboxElementsPage: View {
                 ForEach(elements, id: \.id) { element in
                     NavigationLink {
                         element.pageDescription
-                            .navigationBarMenus()
+                            .navigationBarMenus(title: title)
                     } label: {
                         cardView(for: element)
                     }
@@ -72,7 +71,7 @@ struct DesignToolboxElementsPage: View {
             }
             .padding(.horizontal, 32)
             .padding(.vertical, 20)
-            .navigationBarMenus()
+            .navigationBarMenus(title: title)
             #else
             LazyVGrid(columns: [GridItem(.flexible(), alignment: .topLeading)], spacing: theme.spaces.fixedMedium) {
                 ForEach(elements, id: \.id) { element in
@@ -91,12 +90,12 @@ struct DesignToolboxElementsPage: View {
                     #endif
                 }
             }
-            .oudsGridMargin(.horizontal)
+            .gridMargin(.horizontal)
             .padding(.vertical, theme.spaces.fixedMedium)
-            .navigationBarMenus()
+            .navigationBarMenus(title: title)
             #endif
         }
-        .oudsBackground(theme.colors.bgPrimary)
+        .background(theme.colors.bgPrimary)
         .oudsNavigationTitle(title)
     }
 
@@ -128,13 +127,13 @@ struct DesignToolboxElementsPage: View {
         .background(.regularMaterial, in: .capsule)
         .hoverEffect(.highlight)
         .accessibilityFocused($requestFocus, equals: .some(id: element.id))
-        .oudsRequestAccessibleFocus(_requestFocus, for: .some(id: elements[0].id))
+        .requestAccessibleFocus(_requestFocus, for: .some(id: elements[0].id))
         #else
         Card(
             title: Text(LocalizedStringKey(element.name)),
             illustration: element.illustration)
             .accessibilityFocused($requestFocus, equals: .some(id: element.id))
-            .oudsRequestAccessibleFocus(_requestFocus, for: .some(id: elements[0].id))
+            .requestAccessibleFocus(_requestFocus, for: .some(id: elements[0].id))
         #endif
     }
 }